requirements.txt
django==2.2.1
psycopg2
djangorestframework
drf-nested-routers
django-debug-toolbar
django-flower